+++ src/REvoDesign/REvoDesign.py
@@ -44,7 +44,8 @@ from REvoDesign.phylogenetics import (GREMLIN_Analyser, MutateWorker,
                                       VisualizingWorker)
 from REvoDesign.shortcuts.shortcut_tools import (menu_color_by_plddt,
                                                  menu_dump_sidechains,
-                                                 menu_pssm2csv, menu_real_sc)
+                                                 menu_pssm2csv, menu_real_sc,
+                                                 menu_smiles_conformer_single)
 from REvoDesign.structure import PocketSearcher, SurfaceFinder
 from REvoDesign.tools.customized_widgets import (WorkerThread, decide,
                                                  getExistingDirectory,
@@ -302,6 +303,10 @@ class REvoDesignPlugin(QtWidgets.QWidget):
                 MenuItem(
                     self.bus.ui.actionPSSM_to_CSV,
                     menu_pssm2csv
+                ),
+                MenuItem(
+                    self.bus.ui.actionSMILES_Conformers,
+                    menu_smiles_conformer_single
                 ),
                 MenuItem(
                     self.bus.ui.actionSource_Code,